requirements.txt
flask
langchain
langchain-community
pypdf
python-dotenv
groq
qdrant-client
requests
tqdm
numpy